fix package structure
- repackage by tools (github, jacoco, cobertura)
