+++ src/main/java/com/naver/nid/cover/CoverChecker.java
@@ -17,7 +17,8 @@ package com.naver.nid.cover;
 
 import com.naver.nid.cover.checker.NewCoverageChecker;
 import com.naver.nid.cover.checker.model.NewCoverageCheckReport;
-import com.naver.nid.cover.github.GithubPullRequestManager;
+import com.naver.nid.cover.github.manager.GithubPullRequestManager;
+import com.naver.nid.cover.github.parser.GithubRawDiffReader;
 import com.naver.nid.cover.parser.coverage.CoverageReportParser;
 import com.naver.nid.cover.parser.coverage.model.FileCoverageReport;
 import com.naver.nid.cover.parser.diff.*;
@@ -29,9 +30,7 @@ import com.naver.nid.cover.util.ParameterParser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import java.util.stream.Collectors;
