+++ src/main/java/com/naver/nid/cover/github/manager/model/CommitState.java
@@ -1,0 +1,9 @@
+package com.naver.nid.cover.github.manager.model;
+
+public enum CommitState {
+	SUCCESS, FAILURE, PENDING, ERROR;
+
+	public String toApiValue() {
+		return this.name().toLowerCase();
+	}
+}
