--- src/main/java/com/naver/nid/cover/github/model/CommitState.java
@@ -1,9 +1,0 @@
-package com.naver.nid.cover.github.model;
-
-public enum CommitState {
-	SUCCESS, FAILURE, PENDING, ERROR;
-
-	public String toApiValue() {
-		return this.name().toLowerCase();
-	}
-}
