+++ src/main/java/com/naver/nid/cover/util/ObjectFactory.java
@@ -16,15 +16,15 @@
 package com.naver.nid.cover.util;
 
 import com.naver.nid.cover.checker.NewCoverageChecker;
-import com.naver.nid.cover.github.GithubPullRequestManager;
+import com.naver.nid.cover.github.manager.GithubPullRequestManager;
 import com.naver.nid.cover.parser.coverage.CoverageReportParser;
 import com.naver.nid.cover.parser.coverage.XmlCoverageReportParser;
-import com.naver.nid.cover.parser.coverage.cobertura.CoberturaReportHandler;
-import com.naver.nid.cover.parser.coverage.jacoco.JacocoReportParser;
+import com.naver.nid.cover.cobertura.CoberturaCoverageReportHandler;
+import com.naver.nid.cover.jacoco.JacocoReportParser;
 import com.naver.nid.cover.parser.diff.DiffParser;
 import com.naver.nid.cover.reporter.Reporter;
-import com.naver.nid.cover.reporter.console.ConsoleReporter;
-import com.naver.nid.cover.reporter.github.GithubPullRequestReporter;
+import com.naver.nid.cover.reporter.ConsoleReporter;
+import com.naver.nid.cover.github.reporter.GithubPullRequestReporter;
 
 /**
  * {@link Parameter}에 따라 내부 객체를 생성하는 Factory 객체
@@ -43,7 +43,7 @@ public class ObjectFactory {
 
     public CoverageReportParser getCoverageReportParser() {
         if ("cobertura".equals(param.getCoverageType())) {
-            return new XmlCoverageReportParser(new CoberturaReportHandler());
+            return new XmlCoverageReportParser(new CoberturaCoverageReportHandler());
         } else {
             return new JacocoReportParser();
         }
