+++ src/test/java/com/naver/nid/cover/CoverCheckerTest.java
@@ -3,7 +3,7 @@ package com.naver.nid.cover;
 import com.naver.nid.cover.checker.NewCoverageChecker;
 import com.naver.nid.cover.checker.model.NewCoverageCheckReport;
 import com.naver.nid.cover.checker.model.NewCoveredFile;
-import com.naver.nid.cover.github.GithubPullRequestManager;
+import com.naver.nid.cover.github.manager.GithubPullRequestManager;
 import com.naver.nid.cover.parser.coverage.CoverageReportParser;
 import com.naver.nid.cover.parser.coverage.model.CoverageStatus;
 import com.naver.nid.cover.parser.coverage.model.FileCoverageReport;
@@ -15,7 +15,7 @@ import com.naver.nid.cover.parser.diff.model.DiffSection;
 import com.naver.nid.cover.parser.diff.model.Line;
 import com.naver.nid.cover.parser.diff.model.ModifyType;
 import com.naver.nid.cover.reporter.Reporter;
-import com.naver.nid.cover.reporter.console.ConsoleReporter;
+import com.naver.nid.cover.reporter.ConsoleReporter;
 import com.naver.nid.cover.util.Parameter;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
